posgresql-connection-test/requirements.txt
python-dotenv==1.0.1
cloud-sql-python-connector==1.9.1
pg8000==1.31.2
SQLAlchemy==2.0.24
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3